use postgres as DB + docker

+++ src/structs/sensor_data.rs
@@ -4,13 +4,13 @@ use sqlx::FromRow;
 
 #[derive(Debug, Serialize, Deserialize, FromRow)]
 pub struct SensorData {
-    pub id: u32,
-    pub speed: f32,
+    pub id: i32,
+    pub speed: f64,
     pub created_at: DateTime<Utc>,
 }
 
 impl SensorData {
-    pub const fn new(id: u32, speed: f32, created_at: DateTime<Utc>) -> Self {
+    pub const fn new(id: i32, speed: f64, created_at: DateTime<Utc>) -> Self {
         Self {
             id,
             speed,
